BuscarCart id_cli agora é long ao invez de int

+++ src/carteiraVacinacao/bean/Carteira.java
@@ -193,7 +193,7 @@ public class Carteira {
         }
     }
     
-    public void buscarCart(int id_cli, String nome) {
+    public void buscarCart(long id_cli, String nome) {
         this.carteiraBD.read( id_cli, nome, this );
     }
